Added login controller and service
