+++ src/main/java/peddle/services/LoginService.java
@@ -1,0 +1,10 @@
+package peddle.services;
+
+import org.springframework.http.ResponseEntity;
+import peddle.dto.UserLoginDtoRq;
+
+public interface LoginService {
+
+  ResponseEntity<?> auth(UserLoginDtoRq userLoginDtoRq);
+
+}
